@@ -11,17 +11,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-architects/paper-design": "^1.0.2",
-    "@angular/animations": "14.2.7",
+    "@angular-architects/paper-design": "^1.0.3",
+    "@angular/animations": "15.0.0-rc.0",
     "@angular/cdk": "14.2.5",
-    "@angular/common": "14.2.7",
-    "@angular/compiler": "14.2.7",
-    "@angular/core": "14.2.7",
-    "@angular/forms": "14.2.7",
+    "@angular/common": "15.0.0-rc.0",
+    "@angular/compiler": "15.0.0-rc.0",
+    "@angular/core": "15.0.0-rc.0",
+    "@angular/forms": "15.0.0-rc.0",
     "@angular/material": "14.2.5",
-    "@angular/platform-browser": "14.2.7",
-    "@angular/platform-browser-dynamic": "14.2.7",
-    "@angular/router": "14.2.7",
+    "@angular/platform-browser": "15.0.0-rc.0",
+    "@angular/platform-browser-dynamic": "15.0.0-rc.0",
+    "@angular/router": "15.0.0-rc.0",
     "@ngrx/effects": "^14.3.0",
     "@ngrx/store": "^14.3.0",
     "@ngrx/store-devtools": "^14.3.0",
@@ -32,13 +32,13 @@
   },
   "devDependencies": {
     "@angular-architects/ddd": "^2.1.2",
-    "@angular-devkit/build-angular": "14.2.6",
+    "@angular-devkit/build-angular": "15.0.0-rc.0",
     "@angular-eslint/eslint-plugin": "14.0.4",
     "@angular-eslint/eslint-plugin-template": "14.0.4",
     "@angular-eslint/template-parser": "14.0.4",
     "@angular/cli": "^14.2.0",
-    "@angular/compiler-cli": "14.2.7",
-    "@angular/language-service": "14.2.7",
+    "@angular/compiler-cli": "15.0.0-rc.0",
+    "@angular/language-service": "15.0.0-rc.0",
     "@nrwl/cli": "15.0.0",
     "@nrwl/cypress": "15.0.0",
     "@nrwl/eslint-plugin-nx": "15.0.0",
@@ -67,4 +67,3 @@
     "typescript": "4.8.4"
   }
 }
-
