@@ -1,18 +1,19 @@
 import { LayoutModule } from '@angular/cdk/layout';
-import { HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptors, withLegacyInterceptors } from '@angular/common/http';
 import { enableProdMode, importProvidersFrom } from '@angular/core';
 import { bootstrapApplication } from '@angular/platform-browser';
-import { PreloadAllModules, provideRouter, RouterModule, withDebugTracing, withPreloading } from '@angular/router';
-import { EffectsModule, provideEffects } from '@ngrx/effects';
-import { provideStore, StoreModule } from '@ngrx/store';
-import { provideStoreDevtools, StoreDevtoolsModule } from '@ngrx/store-devtools';
+import { PreloadAllModules, provideRouter, withPreloading } from '@angular/router';
+import { provideEffects } from '@ngrx/effects';
+import { provideStore } from '@ngrx/store';
+import { provideStoreDevtools } from '@ngrx/store-devtools';
 import { reducer } from './app/+state';
 import { AppComponent } from './app/app.component';
 import { APP_ROUTES } from './app/app.routes';
-import { BrowserAnimationsModule, provideAnimations } from '@angular/platform-browser/animations';
+import { provideAnimations } from '@angular/platform-browser/animations';
 
 import { environment } from './environments/environment';
-import { FlightService } from '@nx-example/booking/domain';
+import { authInterceptor } from './utils/auth.interceptor';
+import { LegacyInterceptor } from './utils/legacy.interceptor';
 
 if (environment.production) {
   enableProdMode();
@@ -20,10 +21,17 @@ if (environment.production) {
 
 bootstrapApplication(AppComponent, {
   providers: [
-    importProvidersFrom(HttpClientModule),
+    provideHttpClient(
+      withInterceptors([authInterceptor]),
+      withLegacyInterceptors(),
+    ),
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: LegacyInterceptor,
+      multi: true,
+    },
     provideRouter(APP_ROUTES,
       withPreloading(PreloadAllModules),
-      withDebugTracing()
     ),
     provideStore(reducer),
     provideEffects([]),
